chore: minor cleanups (relationship-metrics module + CLI help + devlog)
P2 hygiene grab-bag from the post-MVP QA gate triage.

- Move computeRelationshipPrecisionRecallF1 out of entity-metrics.ts
  into a new relationship-metrics.ts module so the entity-metrics file
  no longer imports the ExpectedRelationship type. The two files were
  conceptually distinct but bundled together — this is the M3-DIV-010
  cleanup. Updates entity-runner.ts and the eval barrel exports to
  point at the new module.
- Add devlog entry devlog/2026-04-07-graph-step.md for D5 (the M4
  graph step). Closes the documentation gap from P6-DOCS-DEVLOG-01.
- Add Examples blocks via Commander's .addHelpText('after', ...) on
  three top-level CLI commands: mulder ingest, mulder query, and
  mulder pipeline run. The examples cover the most common usage
  patterns and the most useful flag combinations. Closes
  P6-DOCS-CLI-01.
- Update tests/specs/37_qa_error_code_coverage.test.ts: drop the
  PIPELINE_SOURCE_NOT_FOUND, PIPELINE_WRONG_STATUS, PIPELINE_STEP_FAILED,
  and all five EMBED_* codes from the expected-reserved list because
  the previous PR (#117) made them active. The list now matches the
  six codes that genuinely have no thrower today.
